requirements.txt
fastapi
SQLAlchemy
psycopg2-binary
uvicorn
python-multipart
python-dotenv